requirements.txt
kafka-python==2.3.0
lz4==4.4.5
setuptools==80.9.0
wheel==0.45.1
